# cloud/backend/mon-projet-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API Laravel 2026",
        "description": "Documentation de mon API générée avec Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur Local"
        }
    ],
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Lister les rôles",
                "operationId": "ca86f0a206144dfbd695b749b4b28f9d",
                "responses": {
                    "200": {
                        "description": "Liste des rôles"
                    }
                }
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Créer un rôle",
                "operationId": "341b8b162b61267c7bd187500aea5bc4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle",
                                    "niveau"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Administrateur"
                                    },
                                    "niveau": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "is_deleted": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rôle créé"
                    }
                }
            }
        },
        "/api/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Afficher un rôle",
                "operationId": "15c404caf238f1861c81f68205d328e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rôle trouvé"
                    },
                    "404": {
                        "description": "Rôle non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Modifier un rôle",
                "operationId": "c73ab816ae8fb7c34a995d37ba577082",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle",
                                    "niveau"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Utilisateur"
                                    },
                                    "niveau": {
                                        "type": "string",
                                        "example": "2"
                                    },
                                    "is_deleted": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rôle modifié"
                    },
                    "404": {
                        "description": "Rôle non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Supprimer un rôle",
                "operationId": "ffa6ef115d86b00cc6c54a6210019f9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rôle supprimé"
                    }
                }
            }
        },
        "/api/signalements": {
            "get": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Lister les signalements",
                "operationId": "bc5f732d0a6199b18d23b59a8b220d23",
                "responses": {
                    "200": {
                        "description": "Liste des signalements"
                    }
                }
            },
            "post": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Créer un signalement",
                "operationId": "1cef0a98122b3b522b3acde3044deee4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Id_status",
                                    "Id_utilisateur"
                                ],
                                "properties": {
                                    "position_": {
                                        "type": "string",
                                        "example": "-18.8792,47.5079"
                                    },
                                    "commentaire": {
                                        "type": "string",
                                        "example": "Route dégradée"
                                    },
                                    "is_deleted": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "Id_status": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "Id_utilisateur": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Signalement créé"
                    }
                }
            }
        },
        "/api/signalements/{id}": {
            "get": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Afficher un signalement",
                "operationId": "3377083d5e9916548675e39a1fae4652",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Signalement trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Modifier un signalement",
                "operationId": "2d99fa354dfeb952d430f625264f1ff3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Signalement modifié"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Supprimer un signalement",
                "operationId": "fc993b1972ef126f5d52648b5cc781fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Signalement supprimé"
                    }
                }
            }
        },
...
    },
...
}